fix: change imports after moving streamApiLow
